pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-prompt-generation"
version = "0.1.0"
authors = [
    {name = "Sulaiman Mutawalli", email = "sulaimanmutawalli@example.com"},
]
description = "A DSPy-based system for automatically generating optimized prompts"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "dspy-ai>=2.4.0",
]
keywords = ["dspy", "prompt", "generation", "ai", "nlp", "optimization"]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.910",
]

[project.urls]
Homepage = "https://github.com/sulaimanmutawalli/auto-prompt-generation"
"Bug Reports" = "https://github.com/sulaimanmutawalli/auto-prompt-generation/issues"
"Source" = "https://github.com/sulaimanmutawalli/auto-prompt-generation"

[project.scripts]
auto-prompt-gen = "auto_prompt_generation.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["auto_prompt_generation*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
